fix: Solve Attribute Error
Line 189 was throwing:
AttributeError: 'NoneType' object has no attribute 'something'

It seemed that it was attached to the head(head_num) so that was
commented out. There were still issues with 'NoneType' being returned
so then added DF on line 190 to troubleshoot if the DataFrame would
print. This seemed to solve the issue but then was able to comment out
line 190 and uncomment the .head(head_num) in line 189 and the program
went backto functioning properly.